backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
google-genai>=1.29.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
pymongo>=4.8.0
reportlab>=4.2.0
